+++ packages/interfaces-auth-sap/src/auth/IConfig.ts
@@ -2,8 +2,9 @@
  * Configuration - optional composition of authorization and connection configuration
  * Can contain either authorization config, or connection config, or both
  */
+
+import type { IConnectionConfig } from '../auth/IConnectionConfig';
 import type { IAuthorizationConfig } from './IAuthorizationConfig';
-import type { IConnectionConfig } from './IConnectionConfig';
 
 export type IConfig = Partial<IAuthorizationConfig> &
   Partial<IConnectionConfig>;