+++ packages/interfaces-network/src/utils/ITimeoutConfig.ts
@@ -1,5 +1,0 @@
-export interface ITimeoutConfig {
-  default: number;
-  csrf: number;
-  long: number;
-}